NYCCHKBK-2146
git-svn-id: https://omb-dashboardplatform.reisys.com:10039/NYC_Checkbook_2_0/Database@4539 948744eb-f611-4f28-8820-250fb6d8f47d

Former-commit-id: 0c890ad49edf0ef94e3387c8294d1180795d610f

+++ checkbook/DB scripts/ETLScripts/BudgetScripts.sql
@@ -12,7 +12,7 @@ CREATE OR REPLACE FUNCTION etl.updateforeignkeysforbudget(p_load_file_id_in bigi
   RETURNS integer AS $$
 
 DECLARE
-l_count int;
+l_count bigint;
 
 BEGIN
 	/* UPDATING FOREIGN KEY VALUES	FOR BUDGET DATA*/
